@@ -3,21 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@reduxjs/toolkit": "^1.7.1",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
+    "@reduxjs/toolkit": "^1.7.2",
+    "@semantic-ui-react/css-patch": "^1.0.0",
+    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^13.5.0",
-    "@types/gapi": "^0.0.41",
-    "@types/gapi.auth2": "^0.0.56",
     "@types/jest": "^27.4.0",
-    "@types/node": "^16.11.19",
-    "@types/react": "^17.0.38",
+    "@types/node": "^17.0.18",
+    "@types/react": "^17.0.39",
     "@types/react-dom": "^17.0.11",
     "@types/react-redux": "^7.1.22",
-    "@types/react-router": "^5.1.17",
-    "@types/react-router-dom": "^5.3.2",
+    "@types/react-router": "^5.1.18",
+    "@types/react-router-dom": "^5.3.3",
     "@types/redux-persist": "^4.3.1",
-    "axios": "^0.24.0",
+    "axios": "^0.26.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.6",
@@ -25,17 +24,19 @@
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
     "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^2.0.4",
+    "semantic-ui-react": "^2.1.2",
     "styled-components": "^5.3.3",
     "styled-reset": "^4.3.4",
-    "typescript": "^4.5.4",
-    "web-vitals": "^2.1.3"
+    "typescript": "^4.5.5",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "postinstall": "semantic-ui-css-patch",
+    "preinstall": "npx npm-force-resolutions"
   },
   "eslintConfig": {
     "extends": [
@@ -56,10 +57,10 @@
     ]
   },
   "devDependencies": {
-    "@types/styled-components": "^5.1.20",
-    "@typescript-eslint/eslint-plugin": "^5.9.1",
-    "@typescript-eslint/parser": "^5.9.1",
-    "eslint": "^8.6.0",
+    "@types/styled-components": "^5.1.23",
+    "@typescript-eslint/eslint-plugin": "^5.12.0",
+    "@typescript-eslint/parser": "^5.12.0",
+    "eslint": "^8.9.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.25.4",
@@ -67,7 +68,12 @@
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.28.0",
     "eslint-plugin-react-hooks": "^4.3.0",
+    "nth-check": "^2.0.1",
     "prettier": "^2.5.1",
     "redux-persist": "^6.0.0"
+  },
+  "resolutions": {
+    "nth-check": "2.0.1",
+    "postcss": "8.2.13"
   }
 }
