@@ -14,7 +14,7 @@ import {
   InputField,
   InvalidMessage,
   LoginBtn,
-  SocialLoginBtn,
+  SocialLoginBtn2,
   SignupBtn,
 } from './LoginPage.style';
 
@@ -38,15 +38,16 @@ const LoginPage = function LoginPage() {
                 <div>이메일과 비밀번호를 확인해 주세요.</div>
               </InvalidMessage>
               <LoginBtn>로그인</LoginBtn>
-              {/* <SocialLoginBtn> */}
-              <Naver />
-              {/* </SocialLoginBtn> */}
-              {/* <SocialLoginBtn> */}
-              <Google />
-              {/* </SocialLoginBtn> */}
               <SignupBtn>
                 아직 회원이 아니신가요? <p>회원가입 하러가기</p>
               </SignupBtn>
+              <SocialLoginBtn2>
+                <div> SNS를 이용한 소셜 로그인 </div>
+                <div>
+                  <Naver />
+                  <Google />
+                </div>
+              </SocialLoginBtn2>
             </ContentContainer>
           </Container>
         </MainArea>
