@@ -6,6 +6,7 @@ import { CreateUser, DeleteUser, EditUser, FindUserInfo } from '../Services/User
 
 const SignUp = async (req: Request, res: Response) => {
   try {
+    //image 작업필요
     const { nickname, password, image, email }: IUserInput = req.body;
     const findUser = await FindUserInfo(email);
     if (findUser) {
@@ -35,6 +36,7 @@ const SignUp = async (req: Request, res: Response) => {
     return res.status(500).send({ message: 'Internal Server Error', err: err });
   }
 };
+
 const SignOut = async (req: Request, res: Response) => {
   const { email }: IUserInput = req.body;
   const userInfo = await FindUserInfo(email);
@@ -45,17 +47,18 @@ const SignOut = async (req: Request, res: Response) => {
   res.status(200).send({ message: 'Success' });
 };
 
-export const Edit = async (req: Request, res: Response) => {
-  const { nickname, password, image, email }: IUserInput = req.body;
+const Edit = async (req: Request, res: Response) => {
+  //image 작업필요
+  const { nickname, password, image }: IUserInput = req.body;
+  const { email }: any = req.params;
   let userInfo = await FindUserInfo(email);
 
   if (!userInfo) {
     return res.status(404).send({ message: 'Resource Not Found' });
   }
-
+  ///image 코드 확인
   userInfo.nickname = nickname || userInfo.nickname;
   userInfo.password = password || userInfo.password;
-  userInfo.password = email || userInfo.email;
   userInfo.password = image || userInfo.image;
 
   const editUserInfo = await EditUser(userInfo);
@@ -88,7 +91,7 @@ const LogOut = async (req: Request, res: Response) => {
 };
 
 const UserInfo = async (req: Request, res: Response) => {
-  const { email }: IUserInput = req.body;
+  const { email }: any = req.params;
   const userInfo = await FindUserInfo(email);
 
   if (!userInfo) {
