@@ -13,7 +13,7 @@ import { Like } from './Like';
  * @swagger
  *  components:
  *   schemas:
- *    Drink:
+ *    Recipe:
  *      type: object
  *      properties:
  *       id:
@@ -38,8 +38,25 @@ import { Like } from './Like';
  *         type: array
  *         items:
  *          type: string
+ *       likeCount:
+ *         type: number
  *
  *
+ */
+
+/**
+ * @swagger
+ *  components:
+ *   schemas:
+ *    RecipeReturn:
+ *     type: object
+ *     properties:
+ *      data:
+ *       type: array
+ *       items:
+ *         $ref: '#/components/schemas/Recipe'
+ *      message:
+ *       type: string
  */
 
 @Entity()
@@ -53,7 +70,7 @@ export class Drink extends BaseEntity {
   @Column({ type: 'varchar', nullable: true })
   image: string;
 
-  @Column({ type: 'simple-array', nullable: true, default: [] })
+  @Column('text', { array: true, nullable: true, default: '{}' })
   tags: string[];
 
   @Column({ type: 'simple-array', nullable: true })
@@ -64,6 +81,9 @@ export class Drink extends BaseEntity {
 
   @Column({ type: 'varchar', nullable: true })
   Instructions: string;
+
+  @Column({ type: 'int', default: 0 }) // 좋아요 순으로 찾을때 사용
+  likeCount: number;
 
   @CreateDateColumn()
   createdAt: Date;
