@@ -1,8 +1,9 @@
 import { Router } from 'express';
+import { upload } from '../Modules/multer';
 import UserController from '../Controllers/UserController';
-const userRoter = Router();
+const userRouter = Router();
 
-userRoter.post('/signup', UserController.SignUp);
+userRouter.post('/signup', upload.single('image'), UserController.SignUp);
 /**
  * @swagger
  * paths:
@@ -13,9 +14,7 @@ userRoter.post('/signup', UserController.SignUp);
  *    parameters:
  *    - in: body
  *      name: body
- *      content:
- *       application/x-www-form-urlencoded:
- *        schema:
+ *      schema:
  *         type: object
  *         properties:
  *           nickname:
@@ -36,7 +35,7 @@ userRoter.post('/signup', UserController.SignUp);
  *      409:
  *       description: 이미 가입된 이메일
  */
-userRoter.delete('/signout', UserController.SignOut);
+userRouter.delete('/signout', UserController.SignOut);
 /**
  * @swagger
  * paths:
@@ -46,10 +45,8 @@ userRoter.delete('/signout', UserController.SignOut);
  *    summary: 회원탈퇴
  *    parameters:
  *    - in: body
- *      name: body
- *      content:
- *       application/x-www-form-urlencoded:
- *        schema:
+ *      name: email
+ *      schema:
  *         type: object
  *         properties:
  *           email:
@@ -60,7 +57,7 @@ userRoter.delete('/signout', UserController.SignOut);
  *      404:
  *       description: 잘못된 회원 정보
  */
-userRoter.post('/login', UserController.LogIn);
+userRouter.post('/login', UserController.LogIn);
 /**
  * @swagger
  * paths:
@@ -71,9 +68,7 @@ userRoter.post('/login', UserController.LogIn);
  *    parameters:
  *    - in: body
  *      name: body
- *      content:
- *       application/x-www-form-urlencoded:
- *        schema:
+ *      schema:
  *         type: object
  *         properties:
  *           email:
@@ -90,7 +85,7 @@ userRoter.post('/login', UserController.LogIn);
  *      404:
  *       description: 잘못된 회원 정보
  */
-userRoter.post('/logout', UserController.LogOut);
+userRouter.post('/logout', UserController.LogOut);
 /**
  * @swagger
  * paths:
@@ -98,12 +93,11 @@ userRoter.post('/logout', UserController.LogOut);
  *   post:
  *    tags: [User]
  *    summary: 로그아웃
-
  *    responses:
  *      200:
  *       description: 로그아웃 성공
  */
-userRoter.get('/:userEmail', UserController.UserInfo);
+userRouter.get('/:email', UserController.UserInfo);
 /**
  * @swagger
  * paths:
@@ -114,9 +108,7 @@ userRoter.get('/:userEmail', UserController.UserInfo);
  *    parameters:
  *    - in: path
  *      name: email
- *      content:
- *       application/x-www-form-urlencoded:
- *        schema:
+ *      schema:
  *         type: string
  *    responses:
  *      200:
@@ -129,27 +121,27 @@ userRoter.get('/:userEmail', UserController.UserInfo);
  *       description: 잘못된 회원 정보
  */
 
-userRoter.patch('/edit', UserController.Edit);
+userRouter.patch('/:email', upload.single('image'), UserController.Edit);
 /**
  * @swagger
  * paths:
- *  /user/edit:
+ *  /user/{email}:
  *   patch:
  *    tags: [User]
  *    summary: 유저 정보수정
  *    parameters:
+ *    - in: path
+ *      name: userId
+ *      schema:
+ *         type: string
  *    - in: body
  *      name: body
- *      content:
- *       application/x-www-form-urlencoded:
- *        schema:
+ *      schema:
  *         type: object
  *         properties:
  *           nickname:
  *              type: string
  *           password:
- *              type: string
- *           email:
  *              type: string
  *           image:
  *              type: string
@@ -163,4 +155,4 @@ userRoter.patch('/edit', UserController.Edit);
  *      404:
  *       description: 잘못된 회원 정보
  */
-export default userRoter;
+export default userRouter;
