@@ -1,9 +1,13 @@
 import * as express from 'express';
 //import translateRoter from './translate';
-import userRoter from './UserRouter';
+import userRouter from './UserRouter';
+import oauthRouter from './OauthRouter';
+import recipeRouter from './RecipeRouter';
 const app = express();
 
 //app.use('/translate', translateRoter);
-app.use('/user', userRoter);
+app.use('/user', userRouter);
+app.use('/oauth', oauthRouter);
+app.use('/recipe', recipeRouter);
 
 export default app;
